requirements.txt
alembic==1.15.2
attr==0.3.2
asyncpg==0.30.0
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==44.0.2
celery==5.4.0
docutils==0.21.2
fastapi==0.115.12
HTMLParser==0.0.2
httpx==0.28.1
importlib_metadata==8.0.0
ipython==8.12.3
ipywidgets==8.1.6
Jinja2==3.1.2
keyring==25.6.0
lockfile==0.12.2
loguru==0.7.2
mock==5.2.0
Pillow==11.2.1
passlib==1.7.4
protobuf==6.30.2
pydantic==2.11.3
python-slugify==8.0.4
python-multipart==0.0.20
pyOpenSSL==25.0.0
pytest==8.3.5
pytest_asyncio==0.26.0
PyJWT==2.9.0
railroad==0.5.0
Sphinx==8.2.3
SQLAlchemy==2.0.25
thread==2.0.5
tornado==6.4.2
truststore==0.10.1
urllib3_secure_extra==0.1.0
uvicorn==0.34.0